Use only one service with custom @Url for google form and introduce constants class
