+++ app/src/main/java/com/idyndiouck/gads_leaderbord2020/api/ApiConstants.java
@@ -1,0 +1,6 @@
+package com.idyndiouck.gads_leaderbord2020.api;
+
+public class ApiConstants{
+    static String GOOGLE_FORM_ENDPOINT = "https://docs.google.com/forms/d/e/1FAIpQLSf9d1TcNU6zc6KR8bSEM41Z1g1zl35cwZr2xyjIhaMAz8WChQ/formResponse";
+    static String GADS_BASEURL="https://gadsapi.herokuapp.com/api/";
+}
