+++ app/src/main/java/com/idyndiouck/gads_leaderbord2020/api/GadsApiService.java
@@ -4,6 +4,7 @@ import androidx.annotation.NonNull;
 
 import com.idyndiouck.gads_leaderbord2020.model.LearningLeader;
 import com.idyndiouck.gads_leaderbord2020.model.SkillLeader;
+import com.idyndiouck.gads_leaderbord2020.model.Submission;
 
 import java.util.List;
 
@@ -12,12 +13,18 @@ import retrofit2.Callback;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
+import retrofit2.http.Field;
+import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
+import retrofit2.http.POST;
+import retrofit2.http.Url;
+
+import static com.idyndiouck.gads_leaderbord2020.api.ApiConstants.GADS_BASEURL;
 
 public class GadsApiService {
 
     private static RetrofitInterface retrofitInterface = new Retrofit.Builder()
-            .baseUrl("https://gadsapi.herokuapp.com")
+            .baseUrl(GADS_BASEURL)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
             .create(RetrofitInterface.class);
@@ -58,11 +65,43 @@ public class GadsApiService {
                 });
     }
 
-    public static interface RetrofitInterface {
-        @GET("/api/hours")
+    public static void submitProject(Submission submission, @NonNull final ApiResponseCallback<Void> callback) {
+        retrofitInterface.submitProject(ApiConstants.GOOGLE_FORM_ENDPOINT, submission.getFirstName(), submission.getLastName(), submission.getEmail(), submission.getProjectUrl())
+                .enqueue(new Callback<Void>() {
+                    @Override
+                    public void onResponse(Call<Void> call, Response<Void> response) {
+                        if (response.isSuccessful())
+                            callback.onResponse(null);
+                        else
+                            callback.onError(new ApiResponseError(response));
+                    }
+
+                    @Override
+                    public void onFailure(Call<Void> call, Throwable t) {
+                        callback.onError(t);
+                    }
+                });
+
+
+    }
+
+    interface RetrofitInterface {
+        @GET("hours")
         Call<List<LearningLeader>> getLearningLeaders();
 
-        @GET("/api/skilliq")
+        @GET("skilliq")
         Call<List<SkillLeader>> getSkillLeaders();
+
+        @FormUrlEncoded
+        @POST
+        Call<Void> submitProject(
+                @Url String url,
+                @Field("entry.1877115667") String firstName,
+                @Field("entry.2006916086") String lastName,
+                @Field("entry.1824927963") String email,
+                @Field("entry.284483984") String projectUrl
+        );
+
     }
+
 }
