+++ app/src/main/java/com/idyndiouck/gads_leaderbord2020/viewmodel/SubmitViewModel.java
@@ -8,7 +8,7 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.idyndiouck.gads_leaderbord2020.api.ApiResponseCallback;
-import com.idyndiouck.gads_leaderbord2020.api.GoogleFormsApiService;
+import com.idyndiouck.gads_leaderbord2020.api.GadsApiService;
 import com.idyndiouck.gads_leaderbord2020.model.Submission;
 
 public class SubmitViewModel extends ViewModel {
@@ -23,7 +23,7 @@ public class SubmitViewModel extends ViewModel {
                 .postDelayed(new Runnable() {
                     @Override
                     public void run() {
-                        GoogleFormsApiService.submitProject(submission, new ApiResponseCallback<Void>() {
+                        GadsApiService.submitProject(submission, new ApiResponseCallback<Void>() {
                             @Override
                             public void onResponse(Void response) {
                                 status.postValue(STATUS_OK);
